+++ src/test/java/com/nexmo/client/verify/VerifyEndpointTest.java
@@ -1,0 +1,140 @@
+/*
+ * Copyright (c) 2011-2017 Nexmo Inc
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
+package com.nexmo.client.verify;
+
+import org.apache.http.NameValuePair;
+import org.apache.http.client.methods.RequestBuilder;
+import org.apache.http.message.BasicNameValuePair;
+import org.junit.Before;
+import org.junit.Test;
+
+import java.util.HashSet;
+import java.util.List;
+import java.util.Locale;
+import java.util.Set;
+
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertTrue;
+
+public class VerifyEndpointTest {
+
+    private VerifyEndpoint client;
+
+    @Before
+    public void setUp() {
+        client = new VerifyEndpoint(null);
+    }
+
+    @Test
+    public void testConstructVerifyParams() throws Exception {
+        VerifyRequest verifyRequest = new VerifyRequest(
+                "4477990090090",
+                "Brand.com",
+                "Your friend",
+                4,
+                new Locale("en", "GB"),
+                VerifyRequest.LineType.MOBILE
+        );
+
+        RequestBuilder request = client.makeRequest(verifyRequest);
+        List<NameValuePair> params = request.getParameters();
+
+        assertContainsParam(params, "number", "4477990090090");
+        assertContainsParam(params, "brand", "Brand.com");
+        assertContainsParam(params, "sender_id", "Your friend");
+        assertContainsParam(params, "code_length", "4");
+        assertContainsParam(params, "lg", "en-gb");
+        assertContainsParam(params, "require_type", "MOBILE");
+    }
+
+    @Test
+    public void testConstructVerifyParamsMissingValues() throws Exception {
+        VerifyRequest verifyRequest = new VerifyRequest(
+                "4477990090090",
+                "Brand.com",
+                null,
+                0,
+                null,
+                null
+        );
+
+        RequestBuilder request = client.makeRequest(verifyRequest);
+        List<NameValuePair> params = request.getParameters();
+        assertContainsParam(params, "number", "4477990090090");
+        assertContainsParam(params, "brand", "Brand.com");
+        assertParamMissing(params, "code_length");
+        assertParamMissing(params, "lg");
+        assertParamMissing(params, "sender_id");
+        assertParamMissing(params, "require_type");
+        assertParamMissing(params, "country");
+        assertParamMissing(params, "pin_expiry");
+        assertParamMissing(params, "next_event_wait");
+    }
+
+    @Test
+    public void testConstructVerifyParamsWithOptionalValues() throws Exception {
+        VerifyRequest verifyRequest = new VerifyRequest(
+                "4477990090090",
+                "Brand.com"
+        );
+        verifyRequest.setFrom("VERIFICATION");
+        verifyRequest.setLength(6);
+        verifyRequest.setLocale(new Locale("en", "gb"));
+        verifyRequest.setType(VerifyRequest.LineType.LANDLINE);
+        verifyRequest.setCountry("ZZ");
+        verifyRequest.setPinExpiry(60);
+        verifyRequest.setNextEventWait(90);
+
+        RequestBuilder request = client.makeRequest(verifyRequest);
+        List<NameValuePair> params = request.getParameters();
+        assertContainsParam(params, "number", "4477990090090");
+        assertContainsParam(params, "brand", "Brand.com");
+
+        assertContainsParam(params, "code_length", "6");
+        assertContainsParam(params, "sender_id", "VERIFICATION");
+        assertContainsParam(params, "lg", "en-gb");
+        assertContainsParam(params, "require_type", "LANDLINE");
+        assertContainsParam(params, "country", "ZZ");
+        assertContainsParam(params, "pin_expiry", "60");
+        assertContainsParam(params, "next_event_wait", "90");
+
+    }
+
+    private static void assertContainsParam(List<NameValuePair> params, String key, String value) {
+        NameValuePair item = new BasicNameValuePair(key, value);
+        assertTrue(
+                "" + params + " should contain " + item,
+                params.contains(item)
+        );
+    }
+
+    private static void assertParamMissing(List<NameValuePair> params, String key) {
+        Set<String> keys = new HashSet<>();
+        for (NameValuePair pair : params) {
+            keys.add(pair.getName());
+        }
+        assertFalse(
+                "" + params + " should not contain " + key,
+                keys.contains(key)
+        );
+    }
+}